requirements.txt
requests~=2.32.3
litellm~=1.49.6
crewai~=0.80.0
SpeechRecognition~=3.10.4
inflect~=7.4.0
nltk~=3.9.1
langdetect~=1.0.9
transformers~=4.46.1
translate~=3.6.1
pyttsx3~=2.98
html2text~=2024.2.25
pydantic~=2.9.2
embedchain~=0.1.125
langchain-core~=0.3.20
langchain-community~=0.3.7
rpa~=1.50.0
python-dotenv~=1.0.1
langchain~=0.3.7
PyYAML~=6.0.2
autogenstudio~=0.1.5
unstructured~=0.16.5
pywifi~=1.1.12
protobuf~=4.25.5
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.149.0
click~=8.1.7
rich~=13.9.2
tomli~=2.1.0
packaging~=24.1
auth0-python~=4.7.2
cryptography~=43.0.1
tomli_w~=1.1.0
pyvis~=0.3.2
json_repair~=0.25.3
pydantic_core~=2.23.4
chromadb~=0.5.20
appdirs~=1.4.4
regex~=2024.9.11
spacy~=3.8.2
openwakeword~=0.6.0
pycaw~=20240210
colorama~=0.4.6
docker~=7.1.0
psutil~=6.1.0
pyqtgraph~=0.13.7
PyQt5~=5.15.11
numpy~=1.26.4
wikipedia~=1.4.0
onnxruntime~=1.19.2
torch~=2.5.1
sounddevice~=0.5.1
GitPython~=3.1.43
parler_tts~=0.2.1
pyautogen~=0.3.1
tensorboard~=2.18.0
numba~=0.60.0
pytest~=8.3.3
sqlmodel~=0.0.22
setuptools~=75.2.0
fastapi~=0.115.2
loguru~=0.7.2
openai~=1.55.0
pydantic-settings~=2.6.0
typer~=0.9.4
uvicorn~=0.32.0
typing_extensions~=4.12.2
SQLAlchemy~=2.0.35
alembic~=1.13.3
aiofiles~=24.1.0
pillow~=11.0.0
tokentrim~=0.1.13
toml~=0.10.2
tiktoken~=0.7.0
ipython~=8.28.0
html2image~=2.0.5
selenium~=4.25.0
webdriver-manager~=4.0.2
opencv-python~=4.10.0.84
sentence-transformers~=3.3.1
jupyter_client~=8.6.3
ipykernel~=6.29.5
shortuuid~=1.0.13
starlette~=0.40.0
anthropic~=0.39.0
platformdirs~=4.3.6
Send2Trash~=1.8.3
inquirer~=3.4.0
pyreadline3~=3.5.4
prompt_toolkit~=3.0.48